+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the Keyboard Macro Bata extension will be documented in t
 ### [Unreleased]
 - Fix
   - Failure to detect typed characters under some conditions.
+  - Wrong location of multi-cursor after performType under some conditions.
 
 ### [0.1.0] - 2021-11-17
 - Initial release